requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.1.0
xgboost>=2.0.0
lightgbm>=4.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
ccxt>=4.1.0
python-dotenv>=1.0.0
plotly>=5.18.0
dash>=2.14.0
ta>=0.11.0
transformers>=4.36.0
schedule>=1.2.0